fix views

+++ backend/database/migrations/2024_07_01_202820_create_vw_variation_rate_view.php
@@ -12,9 +12,9 @@ class CreateVwVariationRateView extends Migration
      */
     public function up()
     {
-        DB::statement("
+        DB::statement(
+        "
 CREATE OR REPLACE VIEW vw_variation_rate AS
-
     SELECT
         vf1.sai_id,
         vf1.service_id,
@@ -24,19 +24,19 @@ CREATE OR REPLACE VIEW vw_variation_rate AS
         vf1.year AS year1,
         vf2.year AS year2,
         vf1.month,
-        ROUND(vf1.valor_acumulado_agregado, 2) AS total_accumulated_year1,
-        ROUND(vf2.valor_acumulado_agregado, 2) AS total_accumulated_year2,
-        ROUND(vf2.valor_acumulado_agregado - vf1.valor_acumulado_agregado, 2) AS variation_rate_homologous_abs,
+        ROUND(vf1.valor_acumulado_agregado::numeric, 2) AS total_accumulated_year1,
+        ROUND(vf2.valor_acumulado_agregado::numeric, 2) AS total_accumulated_year2,
+        ROUND((vf2.valor_acumulado_agregado - vf1.valor_acumulado_agregado)::numeric, 2) AS variation_rate_homologous_abs,
         CASE
             WHEN vf1.valor_acumulado_agregado = 0 THEN NULL
-            ELSE ROUND(((vf2.valor_acumulado_agregado - vf1.valor_acumulado_agregado) / vf1.valor_acumulado_agregado) * 100, 2)
+            ELSE ROUND(((vf2.valor_acumulado_agregado - vf1.valor_acumulado_agregado) / vf1.valor_acumulado_agregado) * 100::numeric, 2)
         END AS variation_rate_homologous,
-        ROUND(gm1.valor_acumulado_mensal, 2) AS monthly_target_year1,
-        ROUND(gm2.valor_acumulado_mensal, 2) AS monthly_target_year2,
-        ROUND(vf2.valor_acumulado_agregado - gm2.valor_acumulado_mensal, 2) AS variation_rate_contractual_abs,
+        ROUND(gm1.valor_acumulado_mensal::numeric, 2) AS monthly_target_year1,
+        ROUND(gm2.valor_acumulado_mensal::numeric, 2) AS monthly_target_year2,
+        ROUND((vf2.valor_acumulado_agregado - gm2.valor_acumulado_mensal)::numeric, 2) AS variation_rate_contractual_abs,
         CASE
             WHEN gm2.valor_acumulado_mensal = 0 THEN NULL
-            ELSE ROUND(((vf2.valor_acumulado_agregado - gm2.valor_acumulado_mensal) / gm2.valor_acumulado_mensal) * 100, 2)
+            ELSE ROUND(((vf2.valor_acumulado_agregado - gm2.valor_acumulado_mensal) / gm2.valor_acumulado_mensal) * 100::numeric, 2)
         END AS variation_rate_contractual
     FROM
         vw_indicator_accumulated vf1
